extract errors and logger to own files
