+++ src/lib/swankyCommand.ts
@@ -2,44 +2,20 @@ import { Command, Flags, Interfaces } from "@oclif/core";
 import { getSwankyConfig, Spinner } from "./index.js";
 import { SwankyConfig } from "../types/index.js";
 import { writeJSON } from "fs-extra/esm";
-import ModernError from "modern-errors";
-import modernErrorsBugs from "modern-errors-bugs";
-import modernErrorsClean from "modern-errors-clean";
-import modernErrorWinston from "modern-errors-winston";
-import { createLogger, format, transports } from "winston";
-
-export const BaseError = ModernError.subclass("BaseError", {
-  plugins: [modernErrorsBugs, modernErrorsClean, modernErrorWinston],
-});
-
-export const InputError = BaseError.subclass("InputError");
-
-export const UnknownError = BaseError.subclass("UnknownError", {
-  bugs: "https://github.com/swankyhub/swanky-cli/issues",
-});
-
-// const logger = createLogger({
-//   format: ,
-//   transports: [new transports.Http(httpOptions)],
-// })
-
-const winstonCliFormat = format.combine(BaseError.shortFormat(), format.cli());
-const winstonFileFormat = format.combine(BaseError.fullFormat(), format.json());
-const logger = createLogger({
-  format: winstonCliFormat,
-  transports: [new transports.Console()],
-});
-
+import { BaseError, UnknownError } from "./errors.js";
+import { initLogger } from "./logger.js";
+import { Logger } from "winston";
 export type Flags<T extends typeof Command> = Interfaces.InferredFlags<
   (typeof SwankyCommand)["baseFlags"] & T["flags"]
 >;
 export type Args<T extends typeof Command> = Interfaces.InferredArgs<T["args"]>;
 
 export abstract class SwankyCommand<T extends typeof Command> extends Command {
+  static ENSURE_SWANKY_CONFIG = true;
+
   protected spinner!: Spinner;
   protected swankyConfig!: SwankyConfig;
-  static ENSURE_SWANKY_CONFIG = true;
-  static LOG_MODE = "PROD";
+  protected logger!: Logger;
 
   protected flags!: Flags<T>;
   protected args!: Args<T>;
@@ -47,6 +23,7 @@ export abstract class SwankyCommand<T extends typeof Command> extends Command {
   public async init(): Promise<void> {
     await super.init();
     this.spinner = new Spinner();
+    this.logger = initLogger();
 
     const { args, flags } = await this.parse({
       flags: this.ctor.flags,
@@ -56,7 +33,7 @@ export abstract class SwankyCommand<T extends typeof Command> extends Command {
     });
     this.flags = flags as Flags<T>;
     this.args = args as Args<T>;
-    console.log("AAAAAAA", flags);
+
     try {
       this.swankyConfig = await getSwankyConfig();
     } catch (error) {
@@ -77,7 +54,7 @@ export abstract class SwankyCommand<T extends typeof Command> extends Command {
     // add any custom logic to handle errors from the command
     // or simply return the parent class error handling
     const error = BaseError.normalize(err, UnknownError);
-    logger.error(error);
+    this.logger.error(error);
   }
 
   protected async finally(_: Error | undefined): Promise<any> {
